src/iqa/requirements.txt
autologging
python-qpid-proton
amqcfg
dpath
optconstructs
